+++ src/DocumentNumber.Portugal.CitizenCard.Generator/ICitizenCardNumberGenerator.cs
@@ -1,0 +1,8 @@
+﻿namespace DocumentNumber.Portugal.CitizenCard.Generator
+{
+  public interface ICitizenCardNumberGenerator
+  {
+    int CalculateCheckDigit(string uncheckedNumber);
+    string GenerateDocumentNumber();
+  }
+}