+++ src/DocumentNumber.Portugal.Nif.Generator/INifGenerator.cs
@@ -1,0 +1,9 @@
+﻿namespace DocumentNumber.Portugal.Nif.Generator
+{
+  public interface INifGenerator
+  {
+    int CalculateCheckDigit(long uncheckedNumber);
+    string GenerateDocumentNumber();
+    string GenerateDocumentNumber(int startWithNumber);
+  }
+}